requirements.txt
requests
feedparser
schedule
python-dotenv